fix: Refactor frame alpha adjustment functions in km-window
This commit refactors the functions responsible for adjusting the
frame's alpha transparency in km-window.el. It introduces a new helper
function `km-window--add-frame-alpha-param` to consolidate the logic
for modifying frame alpha parameters, including handling of
incompatible properties. Additionally, it extends the functionality to
allow adjusting both the alpha background and the overall frame alpha
transparency with new functions `km-window--frame-inc-alpha` and
`km-window--frame-dec-alpha`. The frame menu transient definitions are
also updated to include options for adjusting the overall frame alpha
transparency.